+++ packages/server/core/src/base/middlewares/customServer/index.ts
@@ -3,8 +3,8 @@ import { ServerHookRunner } from '@core/plugin';
 import { time } from '@modern-js/runtime-utils/time';
 import { Middleware } from '@core/server';
 import { ServerBase } from '@base/serverBase';
-import { createTransformStream } from '../../libs/utils';
-import { ServerReportTimings } from '../../libs/constants';
+import { createTransformStream } from '../../utils';
+import { ServerReportTimings } from '../../constants';
 import { HonoNodeEnv } from '../../adapters/node';
 import {
   createAfterMatchCtx,
