+++ packages/server/core/src/base/middlewares/renderHandler/index.ts
@@ -4,7 +4,7 @@ import { SERVER_DIR } from '@modern-js/utils';
 import { Render } from '@core/render';
 import { Middleware, ServerBaseOptions } from '@core/server';
 import { ServerBase } from '../../serverBase';
-import { checkIsProd, getRuntimeEnv, warmup } from '../../libs/utils';
+import { warmup, checkIsProd, getRuntimeEnv } from '../../utils';
 import { HonoNodeEnv } from '../../adapters/node';
 import { initReporter } from '../monitor';
 import { CustomServer } from '../customServer';
