+++ packages/server/core/src/base/middlewares/renderHandler/ssrCache.ts
@@ -11,8 +11,7 @@ import type {
 import { createMemoryStorage } from '@modern-js/runtime-utils/storer';
 import type { SSRServerContext, ServerRender } from '../../../core/server';
 import { createReadableStreamFromReadable } from '../../adapters/node/polyfills/stream';
-import { createTransformStream } from '../../libs/utils';
-import { getPathname } from '../../libs/request';
+import { createTransformStream, getPathname } from '../../utils';
 
 interface CacheStruct {
   val: string;
