+++ packages/server/core/src/base/utils/error.ts
@@ -1,0 +1,39 @@
+const ERROR_PAGE_TEXT: Record<number, string> = {
+  404: 'This page could not be found.',
+  500: 'Internal Server Error.',
+};
+
+export const createErrorHtml = (status: number) => {
+  const text = ERROR_PAGE_TEXT[status] || '';
+  const title = `${status}: ${text}`;
+  return `<!DOCTYPE html>
+  <html lang="en">
+  <head>
+    <meta charset="utf-8">
+    <meta name="viewport" content="width=device-width">
+    <title>${title}</title>
+    <style>
+      html,body {
+        margin: 0;
+      }
+
+      .page-container {
+        color: #000;
+        background: #fff;
+        height: 100vh;
+        text-align: center;
+        display: flex;
+        flex-direction: column;
+        align-items: center;
+        justify-content: center;
+      }
+    </style>
+  </head>
+  <body>
+    <div class="page-container">
+    <h1>${status}</h1>
+    <div>${text}</div>
+  </body>
+  </html>
+  `;
+};
