+++ packages/server/core/src/base/utils/middlewareCollector.ts
@@ -1,0 +1,32 @@
+export type CollectMiddlewaresResult = {
+  web: any[];
+  api: any[];
+};
+
+export const mergeExtension = (users: any[]) => {
+  const output: any[] = [];
+  return { middleware: output.concat(users) };
+};
+
+export const createMiddlewareCollecter = () => {
+  const webMiddlewares: any[] = [];
+  const apiMiddlewares: any[] = [];
+
+  const addWebMiddleware = (input: any) => {
+    webMiddlewares.push(input);
+  };
+
+  const addAPIMiddleware = (input: any) => {
+    apiMiddlewares.push(input);
+  };
+
+  const getMiddlewares = (): CollectMiddlewaresResult => ({
+    web: webMiddlewares,
+    api: apiMiddlewares,
+  });
+  return {
+    getMiddlewares,
+    addWebMiddleware,
+    addAPIMiddleware,
+  };
+};
