+++ packages/server/core/src/base/utils/serverConfig.ts
@@ -1,8 +1,8 @@
 import * as path from 'path';
 import { compatRequire, fs, DEFAULT_SERVER_CONFIG } from '@modern-js/utils';
 import mergeDeep from 'merge-deep';
+import { ServerConfig } from '@core/plugin';
 import { ServerOptions } from '../../types/config';
-import { ServerConfig } from '../../core/plugin';
 
 export const getServerConfigPath = (
   distDirectory: string,
